Open certificate popover if title box lock clicked
We show two locks in the header bar: the lock in the location entry, and
the lock in the title box. Previously, only the lock in the location
entry was actually functional, but clicking the lock in the title box
ought to work, too, since you should not need to open the location entry
to perform any actions.

Arguably the lock should be removed from the location entry so that this
functionality only exists in one place, but this patch leaves both locks
intact.

https://bugzilla.gnome.org/show_bug.cgi?id=666808