Always update location entry address if unfocused
ephy-location-controller has some suspect code to lock the location
entry, preventing it from being changed when the current web page
changes. This is contrary to the behavior of Firefox, which immediately
updates the location entry when the page changes. I think Firefox's
behavior is less confusing, since it ensures the correct address is
always displayed, but I guess the intent of Epiphany's behavior is to
not change the address if the user is currently typing one.

The problem occurs when the location entry is focused while a new page
is loading: then the old address remains in the location entry, while
the title box subtitle gets updated to the new address. Fix this desync
by always updating the location entry address when the location entry is
unfocused. This means the location entry will only have the incorrect
address as long as it is visible, and will be updated to the correct
address when the title box is shown.

We might additionally want to remove all the locking code so that the
address is always immediately updated.

https://bugzilla.gnome.org/show_bug.cgi?id=734952