ephy-bookmark-properties: close is not an action
We use a traditional X button for close, we don't spell it out as a
suggested-action.

This dialog is a bit confusing because it is used for both creating and
viewing bookmarks. When creating bookmarks, it's an action dialog and we
need to use gtk_dialog_add_button() to create our choices (cancel at the
start of the header bar, and suggested Add at the end). But when not
creating a bookmark, it's a presentation dialog and so calling
gtk_dialog_add_button() is wrong.

https://bugzilla.gnome.org/show_bug.cgi?id=735115