Let the location controller control the title box address
By binding the title box's address directly to the URI property of the
web view, we bypass all our custom logic for displaying the URI,
sometimes leading to desyncs between the URI shown by the location entry
and the URI shown by the title box.

Also, with this change the title of the page now changes at the same
time the displayed address does, whereas previously we would change the
address on the start of load and the title on completion, which looked
odd.

https://bugzilla.gnome.org/show_bug.cgi?id=732713