Fix partially-typed address appearing in subtitle
Now that the EphyLocationController manages the address of the
EphyTitleBox in addition to the EphyLocationEntry, it's no longer safe
to update its address when the address of the location entry changes,
since we want the title box to display the address of the current page,
not whatever the user has typed into the location entry. This is easy to
fix by simply not updating the location controller's address with the
location entry's address.

https://bugzilla.gnome.org/show_bug.cgi?id=732713