ephy-web-view: Fix potential crashers
The "form-auth-data-save-requested" signal was not properly disconnected
as the g_signal_handlers_disconnect_by_func() was using the wrong
object.

The "allow-tls-certificate" signal was not disconnected at all.

The "notify::favicon" was needlessly disconnected (it will be
disconnected when the web view disappears).

And the "cleared" signal for the history service was never disconnected.

g_signal_connect_object() should ensure that all the signals we connect
to that have the web view as data are disconnected when the web view is
destroyed.

https://bugzilla.gnome.org/show_bug.cgi?id=735706