file-helpers: Fix ephy_file_delete_dir_recursively when an error is given
The implementation is checking the given error pointer to check if an
error ocurred, instead of checking the pointer pointed by the passed in
error. This means that it is always failing when an error is passed, but
it works in most of the cases when NULL is passed because the return value
of functions is checked too. There's still a problem when
g_file_enumerator_next_file() returns NULL, because in case of error we
are considering it as end of enumeration. We can simplify the
implementation and the caller code, by using the stdio API instead of
GIO, since we know we are dealing with local files anyway and we are
using the synchronous API.