Fix UI process crash if the web process crashes on start
Currently there is a WebKit bug that causes the web process to crash
when we use webkit_web_view_load_alternate_html() on a local page, say
because Epiphany was closed when a file:// URI was opened in an embed
other than the focused embed. That crash occurs after
web_extension_appeared_cb() but prior to
web_extension_proxy_created_cb(), triggering a call to
web_extension_vanished_cb(), which deletes the EphyWebExtensionProxy
object that is expected to still be alive when
web_extension_proxy_created_cb() is called, crashing the UI process.
We can fix this with a ref/unref pair.

https://bugzilla.gnome.org/show_bug.cgi?id=744896