Do not open an extra overview when starting the browser
If there is no saved session and the user launches Epiphany by clicking
a URI in another application, then we should open only one tab with that
URI. Currently we may open the overview and then the URI in a second tab,
because we do not consider the page to be blank if it is still loading.
Replace this tricky code. If the page is still loading it's not safe to
assume the URI has been set to whatever page was last loaded in it, and we
can't risk losing tabs from a previous session.

https://bugzilla.gnome.org/show_bug.cgi?id=745792