Fix typos and remove a duplicated rule
The rule "#about-subtitle" was duplicated. I assume that the second is the good one because the CSS is read sequentially and the last rule is applied instead the previous rule.

The table of "memory:about" has a typo: solid and transparent properties there were concatenated by mistake.

Removed some white spaces and placed some brackets next to the rule instead below.

https://bugzilla.gnome.org/show_bug.cgi?id=746778