Add context-sensitive menu option to search the web for selected text
When building the context menu in the web process, use the web
extension to find out whether there is text selected and pass it
as custom data to the UI process. Let the UI process use the
selection to build a new menu item that, when activated, will launch
a new tab and search the selected text in the user-preferred search
engine.

https://bugzilla.gnome.org/show_bug.cgi?id=663545