webapp: Open links to the same base domain inside the web app
This fixes the case of web apps like www.foo.com that have a login
button that loads login.foo.com. Since we are just checking the hosts,
they are different, and the login page is opened in a different browser
or profile, making it impossible to login inside the web app. Now, when
the hosts are different we also check the base domain, using
soup_tld_get_base_domain().