Make gettext mandatory and stop using glib-gettext
It's fighting with intltool and intltool is winning, so no point in
keeping it around and making things more complicated. In the future, it
would be good to replace intltool with upstream gettext (not
glib-gettext), but we still need intltool to translate the appdata file.
In the meantime, we'll no longer have two competing applications that
both want to create po/Makefile.in.in.

The AM_GLIB_GNU_GETTEXT macro was defining ENABLE_NLS if gettext is
available, to make i18n optional. But most GNOME programs require
gettext nowadays, and it's installed pretty much everywhere, so there is
no reason for it to be optional anymore. Note, the only reason we were
using glib-gettext before, as far as I see, was to define ENABLE_NLS.