Stop abusing ACLOCAL_AMFLAGS
This is deprecated and it won't work in Automake 2.0, which is
supposedly going to be released later this year. We don't need this
anymore because we call aclocal --install in autogen.sh. Supposedly also
Autoconf 2.70 will be released later this year as well; if that happens,
then autoreconf --install should have learned to pass --install to
aclocal, and we won't need to use it explicitly anymore. In the
meantime, we'll continue to run aclocal in autogen.sh. This is the
recommended practice for all GNOME projects.

There is one downside: libtool 2.4.2 and older will complain that we
have not added -I m4 to ACLOCAL_AMFLAGS. The warning is harmless and
should be ignored. I believe libtool 2.4.3 no longer complains (although
I don't have a new enough libtool to test this).

https://wiki.gnome.org/Projects/GnomeCommon/Migration