Use AX_CODE_COVERAGE instead of GNOME_CODE_COVERAGE
GNOME_CODE_COVERAGE has moved upstream.

Note: This means we now depend on autoconf-archive for builds from git,
which is now a mandatory dependency for GNOME (when building from git).
All macros will be copied into the tarball and distributed
automatically, so there are no new dependencies in our release.

https://wiki.gnome.org/Projects/GnomeCommon/Migration