Reenable -Wmissing-prototype
-Wmissing-prototype is good. Suppressing it for the entire project was a
bad idea.

We'll need to keep -Wno-deprecated-declarations until we can manage to
move away from GtkUIManager, GtkAction, and friends.

I kind of want to remove -Wdeclaration-after-statement as well, because
it is 2015 and it's good to minimize the scope of your variable
declarations by declaring them at point of first use, but this one is
worth discussing before changing.