Silence a deprecation warning coming from egg
Don't want to use the normal compiler warnings here, because I don't
want to fix compiler warnings from egg generally. So we'll just leave
this directory with only the default compiler warnings. But for some
reason, -Wdeprecated-declaration is enabled by default. Maybe we'll
rethink using the default warnings here in the future, but in the
meantime, might as well placate it.

Not upstreamed since the upstream code supports ancient GTK+.

Also note: this code seems to be crashy, but it was crashy BEFORE this
commit too. It seems to be a regression in GTK+ 3.17.