ephy-web-view: Clean up after outliving web extension proxies
If the EphyWebView is destroyed before the EphyWebExtensionProxy
it's attached to does, we'll leave a dangling weak pointer, which
will nullify random memory at the time the web extension proxy
is actually destroyed.

So, prepare for undoing the effects of page_created_cb() in case
we ::dispose() when we still have a web extension attached.

https://bugzilla.gnome.org/show_bug.cgi?id=747422