Don't overwrite page titles in history
When loading a page, we schedule two history updates: the first to reset
the page's title to the empty string (when the web view's title property
is unset after load started) and the second to set the page's title
properly. The first history update is obviously a mistake. It seems like
it should be harmless, so long as the page load is allowed to complete,
but avoiding it seems to solve the bug where pages appear in the history
dialog with URLs for the title, and it's clearly wrong, so let's stop
doing this.

https://bugzilla.gnome.org/show_bug.cgi?id=741385