overview: Move the overview CSS to about.css and generate the HTML from the code
This way we get rid of the overview.html template resource. All the
overview specific classes and ids have been renamed to overview-name,
since the CSS is now in the common abbout.css resource. This fixes the
problems of using a printf template and makes the generation fo the
overview consistent with all other about handlers.

https://bugzilla.gnome.org/show_bug.cgi?id=751385