ephy-location-entry: Get rid of the cell data func to set the text
It seems that changing the model in the cell data func callback confuses
GtkTreeView that keeps validating the rows all the time while the popup
is hidden. The fact that GtkEntry recomputes the size of the popup on
every size_allocate even when the popup is hidden doesn't help either.
So, not changing the model inside the cell data func callback prevents
the rows from being validated indefinitely, but still the cell data func
is called too often in my opinion. And we are always setting the same
text for every row, so I think we should set the title in the model and
get rid of the cell data func.
This patch also ensures that the URL is not shown twice in the
completion menu when there's no title.

https://bugzilla.gnome.org/show_bug.cgi?id=753321