downloads: Add content-type property to EphyDownload
This way it can be monitorized, instead of the caller deciding when to
ask for it during the whole download process. We first try to get it
when the server response is received. Then we try again when the
destination file has been created if the server didn't provide a valid
content type.