Use GdTwoLinesRenderer to display the location entry completion
Accordingly, we get to remove Carlos's horrible hack adding markup to
the completion model (which in MVC should only model, not contain
style), and Claudio's horrible hack adding a bool property to make the
markup optional (to avoid breaking the search provider, which also uses
the completion model).

The markup was originally moved to the completion model in order to
remove the cell data function, which was being run continuously due to
some bug. Because GdTwoLinesRenderer uses real cell attributes to
display the title and url separately, we no longer need to worry at all
about the task it used to perform -- merging the title and url into one
string, with a newline and formatting markup to control size and color
of the second line. GdTwoLinesRenderer takes care of this much more
cleanly.

The only user-visible change is that it is now possible to read URLs in
the completion when the row is selected, as the gray is darker.
Apparently some people can read it fine, but the color was too similar to
the selection blue on the three monitors I tested.

https://bugzilla.gnome.org/show_bug.cgi?id=736364