ephy-session: save and restore web view session state
Using the new WebKit API available in 2.11.3. The session state is
serialized and then encoded as Base-64 to be able to to save it in our
current session state XML file. Now, we always save the session on
close, since there are more things that might have changed in the state
like the scroll position that we don't monitor to schedule saves.

https://bugzilla.gnome.org/show_bug.cgi?id=119432