web-overview: Do not drop thumbnail update requests
When the overview is opened in a new tab, it normally works fine, but
when the overview is opened in the first tab when starting Epiphany,
there are no thumbnails for any of the sites: just empty placeholder
images, at least on my machine. That's not good.

The snapshot service does not have the images ready immediately when
requested by EphyAboutHandler, so EphyAboutHandler is not able to insert
them into the overview HTML when it is created. Instead, it inserts a
placeholder image. Then, when the snapshot service has finished getting
the thumbnail, it makes a D-Bus call to the web extension to convey the
file URI for the thumbnail (for each URL in the overview).
EphyWebOverview replaces the placeholder image with the thumbnail when
it receives the call.

The problem is the call can occur before WebKit has finished loading the
overview HTML, in which case the thumbnail changes get dropped and we
wind up with a bunch of placeholder images. Instead, remember the
thumbnail change requests as they come in, then apply them all after the
document has loaded.

https://bugzilla.gnome.org/show_bug.cgi?id=758470