location-entry: Encode URI before copy/paste
The location entry is currently the only place in our UI where it is
possible to copy a decoded URI. This is basically never what the user
wants.

I am getting tired of accidentally pasting broken URIs like
https://build.webkit.org/results/GTK Linux 64-bit Release
(Tests)/r196216 (13627)/results.html. URIs should be encoded before
stored in the clipboard. Now I can copy
https://build.webkit.org/results/GTK%20Linux%2064-bit%20Release%20(Tests)/r196216%20(13627)/results.html
instead as nature intended.

Of course, if only part of the entry is selected, then it's not a full
URI and should not be messed with; in this case, just copy it raw.

https://bugzilla.gnome.org/show_bug.cgi?id=761660