ephy-session: Save the pending API request URL if EphyWebView address is about:blank
EphyWebView address is NULL (returned as about:blank) between the
load_url() and the WEBKIT_LOAD_STARTED event, but WebKitWebView knows
the pending API request URL. So use that to prevent losing the URL if
the browser is closed before the WEBKIT_LOAD_STARTED event. This usually
happens when the UI process crashes early, right after the session is
loaded but before restored pages have started loading.

https://bugzilla.gnome.org/show_bug.cgi?id=747860