title-box: Do not propagate button press events in title mode
This is a speculative fix for the problem where clicking at just the
right spot in the header bar opens the security popover via the security
indicator in the hidden location entry. I haven't found any good way to
reproduce this problem, unfortunately, but it happens regularly and I
think there's a good chance this will fix it.

This works because there's nothing inside the title box that users
should be allowed to click on when in title mode, except for the
security indicator, which is handled in this very function.

https://bugzilla.gnome.org/show_bug.cgi?id=744307