Activate downloads popover when download is completed
When a download is nearly finished, I find myself watching the download
button, guessing whether the very tip of the download arrow has been
filled in or not, opening the popover to check if it's ready... and
repeating if I guessed wrong. Activate when a download completes to
avoid this.

Not a great final design, it'd be better for the button to glow or
something, but it's a cheap way to tell the user that the download has
finished.

https://bugzilla.gnome.org/show_bug.cgi?id=763141