download: Handle focus stealing prevention properly
Currently, we always pass the start time of the download to
ephy_embed_shell_launch_handler for focus stealing preventing. This is
wrong as it causes apps to not be focused when opening files manually.
For example, I became confused today when I tried to open an image file,
and it opened in eog, but eog appeared beneath the Epiphany window and
did not receive focus, so I could not see it.

Instead, use the current time for focus stealing prevention when the
download action is activated from somewhere external to this file (e.g.
the downloads popover), and only use the start time when the download
action is set in advance via ephy_download_set_download_action(), or
when opening the download automatically without user interaction.

https://bugzilla.gnome.org/show_bug.cgi?id=610586