session: Check restore-session-policy in ephy_session_save()
Instead of doing it in the idle, it's better to do it here before
scheduling the unnecessary idle. Currently it's handled in the idle
because the idle is called manually from ephy_session_close() and this
saves checking it in both places, but now that we need to check it in
ephy_session_close() anyway, there's no value in doing it in the idle.

https://bugzilla.gnome.org/show_bug.cgi?id=762906