session: Fix EPHY_PREFS_RESTORE_SESSION_POLICY_CRASHED
We accidentally broke this when adding support for
WebKitWebViewSessionState. Currently
EPHY_PREFS_RESTORE_SESSION_POLICY_CRASHED is treated exactly the same as
EPHY_PREFS_RESTORE_SESSION_POLICY_ALWAYS. The simplest way to fix this
is to consider that, if we've made it to ephy_session_close(), then we
have not crashed, that's a good place to delete the session file.
Otherwise, that session file had better be saved so it can be loaded if
we crash, ephy_session_close() is the best place to handle this.

https://bugzilla.gnome.org/show_bug.cgi?id=762906