snapshot-service: Take new snapshots after a restart
The previous commit was correct, but the commit message was not. It
stated that removing the erroneous check would cause snapshots to be
taken if the stored snapshot is too old. In fact, removing the check did
not affect behavior at all -- the check was useless -- and so that did
nothing to fix our stale snapshot issue.

If the snapshot path is not in memory cache, this is an indication that
the snapshot is relatively stale (older than the current browsing
session) and should be taken again. But we cannot use the existing
snapshot path memory cache as-is, because it is populated not when
snapshots are taken, but when snapshots are looked up. Augment the cache
to remember which paths were snapshotted since Epiphany was last
restarted, and check it to determine whether we're due for a snapshot of
the current page. Also, remove use of the cache from EphyWebView, since
that would prevent snapshots from ever being taken.

This really fixes the issue.

https://bugzilla.gnome.org/show_bug.cgi?id=763184