snapshot-service: Remove unused functions to retrieve pixbufs
We have a bunch of complicated code to support retrieving GdkPixbufs
from the snapshot service, but Epiphany does not use these and I don't
expect it to in the future. The snapshot service is too complex right
now and it's more important to simplify than to keep dead code around.

https://bugzilla.gnome.org/show_bug.cgi?id=765863