Allow running system web apps outside GNOME
Expect system web apps to pass the desktop file path via the command
line using the existing --application-mode flag, rather than requiring
that Epiphany be started via Gio. This allows Epiphany to compete with
Electron as a choice for running web apps that work outside GNOME. It
also breaks any existing system web apps, but that's probably fine as
this feature is new in 3.20 and I'm not aware of anybody using it yet.

Note this "API" is subject to change prior to the 3.22.0 release.

https://bugzilla.gnome.org/show_bug.cgi?id=767101