sync-service: Process storage requests in the same order they were sent
Since all storage requests are made asynchronously (thus the exact
moment when the responses arrive can't be known) and since implementing
the sync logic requires multiple consecutive storage requests to be
made, a certain mechanism is needed to ensure that the responses are
processed in the right order, one after another, to avoid conflicts.

To achieve this, use a queue to hold new requests and a flag to tell
whether there is currently another request in progress: if there is
currently another request being transmitted, then the new one has to
wait in the queue, otherwise, it is free to go. With this, it becomes
the responsibility of the current request to release the next one
waiting in the queue once the response has been processed. This is done
by calling ephy_sync_service_release_next_storage_message() at the end
of every callback that handles a response from the Storage Server.