Add URI handler for view source mode
This brings back the view source handler that lived in master in late 2016
and early 2017, so we don't have to open page source in gedit anymore.

It's not perfect, in that it could get the source code from the wrong
tab if two different tabs are displaying the same URI, but let not
perfect be the enemy of the good. Content will be fetched from the
network if no tab is displaying the desired URI.

No syntax highlighting because I never found a library that handles this
efficiently enough for our needs.

https://bugzilla.gnome.org/show_bug.cgi?id=738475