Open ftp URIs in external application
Likely we're never going to support FTP, at least not unless a wild
volunteer appears to implement it.

So:

First, try the default URI handler. It might be Firefox, which is odd
but will work. Whatever gvfs-open picks on my computer fails with an
error that the specified location is not mounted, so forcibly fall back
to Nautilus if the default application did not work.