view-source-handler: Get content from existing embed if possible
Instead of always hitting the network to download the page again, first
try to get the source from an existing web view that matches the URI to
be displayed. Be careful to ignore the web view if it has a deferred
load or hasn't finished loading yet.

It's not perfect, in that it could get the source code from the wrong
tab if two different tabs are displaying the same URI, but let not
perfect be the enemy of the good.

https://bugzilla.gnome.org/show_bug.cgi?id=738475