title-box: Fix clicking on security indicator
It doesn't work because EphyTitleBox is now a GtkBox instead of a
GtkStack, so it no longer has a GdkWindow and cannot receive events
anymore. We need to use a GtkEventBox. The simplest solution is to pack
just the GtkImage into the event box, but I had trouble positioning the
popover properly with that strategy. Next easiest solution is to turn
the EphyTitleBox itself into a GtkEventBox. This works nicely.

https://bugzilla.gnome.org/show_bug.cgi?id=771229