Create only an EphyLocationEntry *or* an EphyTitleBox
Now that EphyTitleBox no longer switches back and forth between location
entry mode and title mode... instead of wastefully creating both an
EphyLocationEntry and an EphyTitleBox and only ever using one or the
other, only create the one that's needed: an EphyLocationEntry normally,
but an EphyTitleBox if we're in app mode.

This requires moving the lock-clicked signal of EphyLocationEntry and
EphyTitleBox up to the EphyTitleWidget interface, and adjusting them to
match (carry the GdkRectangle at which to position the
EphySecurityPopover).