Improve overview layout in small windows
It looks pretty bad currently if you resize Epiphany down to the point
where five snapshots no longer fit in a row; then we wind up with four
rows of snapshots, two with four snapshots and two with one snapshot.
This sucks. Adrian found a great way to avoid this with CSS magic.