Add an option to never remember passwords
Some users want to never remember passwords on particular websites. Add
an option to the info bar to ensure it never appears in the future for a
given website.

Note there is an existing option to disable password saving on all
sites in the preferences dialog.

Also note that this requires moving EphyHostsManager to lib/ to allow it
to be used from the web extension.

Currently there is no UI to undo the decision if the user accidentally
clicks Never. It's possible to fix by editing hosts.ini, but ideally we
would add UI for this somewhere. (The same UI is needed to manage
notification permissions.)

https://bugzilla.gnome.org/show_bug.cgi?id=612988