embed: don't attempt to parent notification container twice
If the EphyNotificationContainer is already contained by another embed
(including, in particular, another embed in another window) then don't
try to parent it again. Similarly, don't attempt to unparent an
EphyNotificationContainer unless it's really contained by this embed.

This should probably be cleaned up better, but it's not simple to reason
about what should happen to the EphyNotificationContainer when there
are multiple windows.