sync: Fix endless sign-in confirmation loop
The initial approach of sign-in was a faulty, since it didn't take into
consideration the "sign-in confirmation email" feature that Mozilla was
going to introduce.

Instead of destroying the session and reloading the Firefox Accounts iframe
in case of an unverified account, keep the session alive and, with the
help of keyFetchToken, poll the '/account/keys' endpoint repeatedly until
the verification had finished, without reloading the iframe. When a 200 OK
response is received from the server, proceed to derive the sync keys
and display the 'signed in as' panel.