Revert "webapp: Open links to the same base domain inside the web app"
This reverts commit 0b4464039e0e1175abd0f90e2700631991d643c0.

I don't know what website this change was made for, nor do I have a good
solution to make web apps work with that unfortunate website, but I have
my own web app where this behavior is confusing and unexpected.
Different domains are treated as completely unrelated by web browsers,
and web apps should not be any exception. At any rate, this is the wrong
place for the check too, as it results in URLs with different origins
being reported to have the same origin.

I really don't know if the original problem can be fixed properly or
not, but hopefully the login page link would be opened in a related view
and a heuristic could be developed based on that.