main: Remove --netbank-mode
It's the same as incognito-mode, and it's confusing. The two people
in the world using this flag can update their scripts.

Besides, do you really think it's a good idea to choose a memorable
password for your bank? Using the password manager is arguably much more
important here than elsewhere.