Remove the enable-javascript preference
Unfortunately too many Epiphany features are implemented with JavaScript
message handlers. We never noticed that they were broken for ages, which
indicates that nobody is using this setting. It's too coarse-grained.
hardly anybody wants to browse the web with no JavaScript at all;
rather, people want NoScript-style ability to block scripts only on
particular sites. We can bring this back in the future if it's (a)
implemented on a host-specific basis, (b) preferably with UI so that
people can actually use it, our little NoScript, and (c) accordingly,
not going to break our script message handlers.