Move EphyUriTester to the UI process and make it a D-Bus interface
For one thing, it's silly to have n different EphyUriTesters in n
different web processes, each one loading up adblock filters separately.
Once upon a time, this used to cause big problems when the different web
processes would stomp on the global filters file. For us to ever
implement a real filters configuration dialog, this needs to be handled
in one place, by the UI process, and web extensions must just query the
UI process when they want to use the URI tester.

For another: it's basically required to use libhttpseverywhere
effectively. We're going to have to set up almost this same exact
interface for libhttpseverywhere, so might as well do it for adblock
too. Why is it needed? Because loading HTTPS Everywhere rulesets takes
~2 seconds apiece. It's much too long to do each time we open a new
browser tab, so we should do it in the UI process instead.

An unfortunate consequence of this is that our GDBusConnection use in
the web process can no longer be asynchronous. This is because we must
have an EphyUriTesterProxy completely ready to be used before the first
URI request is ready. We must handle each URI request synchronously as a
consequence of WebKit's signal-based API. What a shame!