web-overview: Don't ever duplicate thumbnails
When we were restored from page cache, we wound up with a page that has
a populated group of overview thumbnails in the DOM, but EphyWebOverview
itself has no knowledge of them, because the document-loaded signal
never gets emitted in this case (well, actually it gets emitted when the
document is loaded the FIRST time)....

Fix this by removing all the code that's designed to manually add items
to the overview. I don't see why it exists; EphyAboutHandler should
always take care of adding overview items for us, so we shouldn't need
to do any DOM manipulation here. Then manually call
ephy_web_overview_document_loaded so that we populate our list of
overview items.

(If you're paying extra close attention, you'll notice that this commit
deletes all the code modified in the previous commit.)