history-service: Initialize tables even in read-only mode
The history service just is not built to handle the case where no
history tables exist. We either need to upgrade every error path in this
directory to not call g_warning in read-only mode, or just create the
tables. Creating tables is harmless; we just don't want to insert
history into them.