main: default to the default profile in incognito mode
epiphany --incognito-mode unexpectedly behaves differently than the
incognito mode app menu option. It is supposed to steal data from the
default profile directory, but this is currently broken unless the
profile is passed explicitly. The problem is that ephy_dot_dir() is
guaranteed to be NULL here, because ephy_file_helpers_init() has not yet
been called.