bookmarks: Split list model functionality into separate class
This is much less efficient, but it also makes the code much simpler and
more robust. Using EphyBookmarksManager as the list model was a good
choice at the time, when we were not thinking about smart bookmarks, but
now that we need to handle both it doesn't work so well. My initial
implementation was complex and buggy. Simplify this.