Improve handling of search the web menu item
We shouldn't be putting the entire context menu item label into the
GAction parameter. This is excessive and error-prone. We've started
crashing after changing the quotation characters from ' to “”, for
example. Also, some languages need to replace the quotation characters
in translations, making the previous behavior especially risky. Lastly,
extracting the search text from the ellipsized label means our search
text itself will be ellipsized, which is undesirable and not the
previous behavior prior to the GAction port.

https://bugzilla.gnome.org/show_bug.cgi?id=774318