Use ISO_CODES_PREFIX to find localedir of iso-codes
This commit makes it possible to find translation data of iso-codes when
epiphany and iso-codes are installed in different prefixes, which is
common in JHBuild environments.

We hardcode share/locale here because iso-codes.pc doesn't have the path
of localedir written in it and most users and distributions don't modify
the path of localedir.

https://bugzilla.gnome.org/show_bug.cgi?id=774464