window: fix web view receiving events twice
The current code propagates the event to the web view, then chains up if
the web view doesn't handle the event. But chaining up causes GtkWindow
to propagate the event to the web view yet again. Surely we never want
to do that, so stop doing it.

I think there must be some other bug here, though, in WebKit, that
causes WebKit to sometimes do something with the event but then
propagate anyway, which is wrong. If I'm right, then WebKit is
unfortunately still broken, but this works around it in Epiphany and is
the right thing to do anyway, since sending the same event to the web
view twice is nonsense regardless of whether the web view propagates it
or not.

https://bugzilla.gnome.org/show_bug.cgi?id=764653