Fix bookmarks migration
EphyBookmarksManager always creates an empty bookmarks.gvdb on its init
method, and migrate_bookmarks() checks whether bookmarks.gvdb already
exists to avoid migrating the bookmarks twice. Since migrate_bookmarks()
creates the EphyBookmarksManager before checking for bookmarks.gvdb, it
always exists and bookmarks are never migrated. This is a regression
introduced in commit 54006df8dba31a5a58686e7d2473d64f3e454683.
This patch creates the manager after the initial checks, so that we also
avoid creating the manager and all the sync IO involved if we are not
going to use it.

https://bugzilla.gnome.org/show_bug.cgi?id=774785