uri-tester: Fix criticals before libhttpseverywhere is initialized
I added a performance optimization to short-circuit the deferred request
system for non-http:// URIs, since there's no reason to wait for
libhttpseverywhere to be ready before processing them. But I forgot to
actually not use libhttpseverywhere in that case, and changed
libhttpseverywhere to emit the critical immediately instead of after
checking the URI scheme.